Rename utf-8''C1M2_Assignment.ipynb to Evaluation of Diagnostic Models
In this assignment, we will be working with the results of the X-ray classification model we developed in the previous assignment. In order to make the data processing a bit more manageable, we will be working with a subset of our training, and validation datasets. We will also use our manually labeled test dataset of 420 X-rays.